refactor(BridgeGameController): 기존 컨트롤러 병합 및 리팩터링
