+++ __tests__/GameControllerTest.js
@@ -1,4 +1,4 @@
-const { isCommandRetry } = require('../src/GameController');
+const { isCommandRetry } = require('../src/BridgeGame');
 
 describe('BridgeGame 클래스 테스트', () => {
   test('IsCommandRetry - 커맨드에 맞게 올바른 값을 반환하는지 검사', () => {
