+++ src/BridgeGameController.js
@@ -1,23 +1,20 @@
 const { Console } = require('@woowacourse/mission-utils');
-const {
-  initBridgeGame,
-  selectWrongDirection,
-  canMoveNext,
-  isCommandRetry,
-  isWrongDirection,
-} = require('./GameController');
+const { isWrongDirection, isCommandRetry } = require('./BridgeGame');
+const BridgeGame = require('./BridgeGame');
+const BridgeMaker = require('./BridgeMaker');
+const { generate } = require('./BridgeRandomNumberGenerator');
 const {
   readMoving,
-  readBridgeSize,
   repeatReadMoving,
   readGameCommand,
+  readBridgeSize,
 } = require('./InputView');
 const { printResult, printMap } = require('./OutputView');
 
 class BridgeGameController {
   static readSize(size) {
     try {
-      const game = initBridgeGame(size);
+      const game = BridgeGameController.initBridgeGame(size);
       readMoving(game, BridgeGameController.moveNext);
     } catch (err) {
       Console.print(err);
@@ -25,25 +22,42 @@ class BridgeGameController {
     }
   }
 
+  static initBridgeGame(size) {
+    const bridge = BridgeMaker.makeBridge(size, generate);
+    console.log(bridge);
+    const game = new BridgeGame(bridge);
+    return game;
+  }
+
   static moveNext(game, direction) {
     try {
       game.move(direction);
-      if (game.isEndOfBridge()) {
-        printResult(game);
-      } else if (isWrongDirection(game)) {
-        printMap(game);
-        readGameCommand(game, BridgeGameController.retryDoing);
-      } else if (canMoveNext(game)) {
-        printMap(game);
-        repeatReadMoving(game, BridgeGameController.moveNext);
-      }
+      BridgeGameController.checkGameProcess(game);
     } catch (err) {
       Console.print(err);
       repeatReadMoving(game, BridgeGameController.moveNext);
     }
   }
 
-  static retryDoing(game, command) {
+  static checkGameProcess(game) {
+    if (isWrongDirection(game)) {
+      printMap(game);
+      readGameCommand(game, BridgeGameController.checkRetry);
+    } else if (game.isEndOfBridge()) {
+      printResult(game);
+    } else if (BridgeGameController.canMoveNext(game)) {
+      printMap(game);
+      repeatReadMoving(game, BridgeGameController.moveNext);
+    }
+  }
+
+  static canMoveNext(game) {
+    if (!game.isEndOfBridge()) return true;
+
+    return false;
+  }
+
+  static checkRetry(game, command) {
     try {
       if (isCommandRetry(command)) {
         game.retry();
@@ -51,7 +65,7 @@ class BridgeGameController {
       } else printResult(game);
     } catch (err) {
       Console.print(err);
-      readGameCommand(game, BridgeGameController.retryDoing);
+      readGameCommand(game, BridgeGameController.checkRetry);
     }
   }
 }
